+++ dataquality/clients/__init__.py
@@ -1,4 +1,4 @@
-from dataquality.clients.http_client import api_client
+from dataquality.clients.api import api_client
 from dataquality.clients.objectstore import object_store
 
 __all__ = ["api_client", "object_store"]
